requirements.txt
# worker/requirements.txt

fastapi==0.109.1
uvicorn[standard]==0.27.0
pydantic>=2.0.0
python-dotenv==1.0.0

boto3==1.34.0
sqlalchemy==2.0.40
pymysql==1.1.1
requests==2.31.0
cryptography

chardet